% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tcrm.R
\name{add_ccl_tcrm}
\alias{add_ccl_tcrm}
\title{Add CCL TCRM to TCRM Data}
\usage{
add_ccl_tcrm(dt, tcrm_column)
}
\arguments{
\item{dt}{A data.table containing TCRM data. It must have a "date" column and a column matching tcrm_column.}

\item{tcrm_column}{A character string specifying the name of the column in dt to be multiplied by tx_gap.}
}
\value{
The original data.table dt with an additional column named <tcrm_column>_ccl.
}
\description{
Merges a data.table with TX Gap data and adds a new column that multiplies the specified tcrm column by the tx_gap.
}
\examples{
\dontrun{
  # Suppose dt has a "date" column and a "value" column:
  dt <- data.table(date = as.Date('2020-01-01') + 0:4, value = 1:5)
  dt <- add_ccl_tcrm(dt, "value")
}
}
