% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_by_inflation.R
\name{adjust_by_inflation}
\alias{adjust_by_inflation}
\title{Adjust price columns by inflation}
\usage{
adjust_by_inflation(
  dt,
  inflation_dt,
  date_column = "date",
  price_columns = c("lastPrice", "openPrice", "maxPrice", "minPrice")
)
}
\arguments{
\item{dt}{A data.table where each row contains one or more prices on a certain date.}

\item{inflation_dt}{A data.table a \code{date} column containing the first day of each month,
and an \code{inflation_ajust_coef} column with the inflation adjustment coefficient for that month,
as produced by the \code{calc_inflation_from_cpi} function.}

\item{date_column}{A String with the name of the date column in \code{dt}.}

\item{price_columns}{A vector of Strings with the names of the price columns in dt to be adjusted.}
}
\value{
A data.table with the same columns as \code{dt}, plus the adjusted price columns.
}
\description{
Takes a data.table with prices and a data.table with inflation coefficients
(such as the one produced by the \code{calc_inflation_from_cpi} function) and returns a data.table
with the same columns as the original one, plus the adjusted price columns.

Prices for dates after the last inflation coefficient are not adjusted.
}
