% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_inflation_from_cpi.R
\name{calc_inflation_from_cpi}
\alias{calc_inflation_from_cpi}
\title{Calculate inflation from CPI DataFrame}
\usage{
calc_inflation_from_cpi(cpi_dt)
}
\arguments{
\item{cpi_dt}{A data.table with monthly CPI data. Must have a \code{date} and a \code{cpi} column.}
}
\value{
The same data table as the input, with the following columns added:
\itemize{
\item \code{monthly_inflation}: the monthly inflation for that month.
\item \code{annual_inflation}: the yearly inflation between that month and the same month of the previous year.
\item \code{inflation_ajust_coef}: a coefficient to multiply prices of that month to adjust them for inflation
and translate them to current prices.
}
}
\description{
Takes a data.table with monthly CPI data and calculates monthly and annual inflation.
Also adds a column with the inflation adjustment coefficient, which is the ratio between the last CPI and the current CPI.
}
