% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpi_us.R
\name{get_cpi_us}
\alias{get_cpi_us}
\title{Get monthly Consumer Price Index data from FRED}
\usage{
get_cpi_us()
}
\value{
A data table with monthly CPI data.
}
\description{
Wrapper for getSymbols from quantmod to download monthly CPI data from FRED.
}
