% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tcrm.R
\name{get_tcrm}
\alias{get_tcrm}
\title{Get TCRM Data from a URL}
\usage{
get_tcrm(local_dir = "/tmp")
}
\arguments{
\item{local_dir}{A character string specifying the local directory to save the downloaded file. Defaults to "/tmp".}
}
\value{
A data.table containing the processed TCRM data.
}
\description{
Downloads an Excel file from the provided URL, processes the data by skipping the first row, converts the first column to Date type, and returns the data as a data.table.
}
\examples{
\dontrun{
  tcrm_data <- get_tcrm()
}
}
