% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tx_a3500.R
\name{get_tx_a3500}
\alias{get_tx_a3500}
\title{Get TX A3500 Data from a URL}
\usage{
get_tx_a3500(local_dir = "/tmp")
}
\arguments{
\item{local_dir}{A character string specifying the local directory to save the downloaded file. Defaults to "/tmp".}
}
\value{
A data.table containing the processed TX A3500 data.
}
\description{
Downloads an Excel file from the provided URL, reads data from headers on row 4, extracts only the second and third columns, converts the first extracted column to Date type (using format "yy-mm-dd"), renames the columns to "date" and "tc_a3500", and returns the data as a data.table.
}
\examples{
\dontrun{
  tc_a3500_data <- get_tc_a3500()
}
}
