% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tx_gap.R
\name{get_tx_gap}
\alias{get_tx_gap}
\title{Get TX Gap Data}
\usage{
get_tx_gap()
}
\value{
A data.table containing the joined data with TC A3500 and CCL data.
}
\description{
Loads CCL data using \code{load_ccl()} and TC A3500 data using \code{get_tc_a3500()}, then performs a rolling join on the "date" column. The CCL data is subset to include only the date and the rate (renamed as "ccl") before joining.
}
\examples{
\dontrun{
  tx_gap <- get_tx_gap()
}
}
