% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ccl.R
\name{load_ccl}
\alias{load_ccl}
\title{Load historic CCL data persisted by broker-management}
\usage{
load_ccl(ccl_id = "cclGGAL")
}
\arguments{
\item{ccl_id}{A string with the filename of the CSV to be loaded.}
}
\value{
A data frame with historic exchange rates.
}
\description{
Loads a file with historic CCL data from the currencies directory in the broker-management project.
The historic values have been obtained as the ratio between the price of a certain stock and the price of it's GDR.
}
