% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_market_data.R
\name{load_market_data}
\alias{load_market_data}
\title{Load historic market data persisted by broker-management}
\usage{
load_market_data(ticker, exchange = c("BCBA", "NYSE", "LON", "NASDAQ"))
}
\arguments{
\item{ticker}{A string with the ticker of the asset to be loaded.}

\item{exchange}{A string identifying the stock exchange where the asset is traded.}
}
\value{
A data frame with historic market data.
}
\description{
Loads a file with historic market data from the market directory in the broker-management project.
}
