% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_my_data.R
\name{load_my_data}
\alias{load_my_data}
\title{Load certain datasets persisted by broker-management}
\usage{
load_my_data(
  data_id = c("current_position", "recent_prices", "unified_movements", "unified_orders")
)
}
\arguments{
\item{data_id}{A string identifying the dataset to be loaded.
Can be one of "current_position", "recent_prices", "unified_movements", "unified_orders"}
}
\value{
A data frame
}
\description{
Factory method for loading a dataframe from the outputs directory in the broker-management project
}
\examples{
load_my_data("current_position")
load_my_data("recent_prices")
load_my_data("unified_movements")
load_my_data("unified_orders")
}
