% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data_to_ccl.R
\name{market_data_to_ccl}
\alias{market_data_to_ccl}
\title{Convert market data prices to CCL}
\usage{
market_data_to_ccl(df_market, df_exch_rate)
}
\arguments{
\item{df_market}{A data.table with the market data, as returned by \code{load_market_data()}}

\item{df_exch_rate}{A data.table with the exchange rate data, as returned by `load_ccl()}
}
\value{
The same data.table with it's prices converted to the new currency.
}
\description{
Use this function to convert a data.table with market data to CCL currency.
If there is no exchange rate for a certain date, the original price is lost.
}
