% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_factory.R
\name{path_factory}
\alias{path_factory}
\title{Get path to a folder in the broker-management project}
\usage{
path_factory(dir_id = c("resources", "outputs", "currencies", "market"))
}
\arguments{
\item{dir_id}{A string identifying the directory to get the path to.
Can be one of "resources", "outputs", "currencies", "market".}
}
\value{
A string with the path to the folder
}
\description{
Factory method for getting a string with the path to a folder in the broker-management project
}
\examples{
path_factory("resources")
path_factory("outputs")
path_factory("currencies")
path_factory("market")
}
